+++ src/BDSBunchSigmaMatrix.cc
@@ -28,7 +28,8 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "CLHEP/RandomObjects/RandMultiGauss.h"
 #include "CLHEP/Units/PhysicalConstants.h"
 
-BDSBunchSigmaMatrix::BDSBunchSigmaMatrix()
+BDSBunchSigmaMatrix::BDSBunchSigmaMatrix():
+  BDSBunchGaussian("sigma matrix")
 {;}
 
 void BDSBunchSigmaMatrix::SetOptions(const BDSParticleDefinition* beamParticle,