+++ src/BDSBunchUserFile.cc
@@ -42,6 +42,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 template <class T>
 BDSBunchUserFile<T>::BDSBunchUserFile():
+  BDSBunch("user file"),
   distrFile(""),
   distrFilePath(""),
   bunchFormat(""),