+++ src/BDSIMClass.cc
@@ -252,6 +252,7 @@ int BDSIM::Initialise()
 					  globalConstants->BeamlineTransform(),
 					  globalConstants->BeamlineS(),
 					  globalConstants->GeneratePrimariesOnly());
+  G4cout << "Bunch distribution: " << bdsBunch->Name() << G4endl;
   /// We no longer need beamParticle so delete it to avoid confusion. The definition is
   /// held inside bdsBunch (can be updated dynamically).
   delete beamParticle;